restructure for instance compiler, add ast package
