+++ src/main/java/org/clafer/instance/InstanceModel.java
@@ -5,10 +5,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import org.clafer.ast.AstClafer;
 import org.clafer.ast.AstConcreteClafer;
 import org.clafer.common.Check;
-import org.sysml.*;
+import org.sysml.ast.SysmlProperty;
+import org.sysml.compiler.InstanceSysmlCompiler;
 import org.sysml.pprinter.SysmlPrinter;
 
 /**
@@ -53,15 +53,6 @@ public class InstanceModel {
         return typedTopClafer;
     }
 
-    private SysmlProperty compileInstance(InstanceClafer model) {
-        String propertyName = model.getType().getName();
-        ArrayList<SysmlBlockDefElement> children = new ArrayList<SysmlBlockDefElement>();
-        for (InstanceClafer child : model.getChildren()) {
-            children.add(compileInstance(child));
-        }
-        return new SysmlProperty(new SysmlBlockVisibility(SysmlVisibilityOption.PLUS), new SysmlPropertyType("part"), propertyName, children.toArray(new SysmlBlockDefElement[children.size()]));
-    }
-
     /**
      * Print solution as SysMLv2
      *
@@ -74,7 +65,8 @@ public class InstanceModel {
     public void printSysml(Appendable out) throws IOException {
         for (InstanceClafer top : topClafers) {
             SysmlPrinter pprinter = new SysmlPrinter(out);
-            SysmlProperty model = compileInstance(top);
+            InstanceSysmlCompiler compiler = new InstanceSysmlCompiler();
+            SysmlProperty model = compiler.compile(top);
             pprinter.visit(model, "");
         }
     }
