+++ src/main/java/org/sysml/ast/SysmlBlockDefElement.java
@@ -1,4 +1,4 @@
-package org.sysml;
+package org.sysml.ast;
 
 public interface SysmlBlockDefElement extends SysmlExpr, SysmlId {
 }
