+++ src/main/java/org/sysml/ast/SysmlBlockVisibility.java
@@ -1,4 +1,4 @@
-package org.sysml;
+package org.sysml.ast;
 
 /**
  * Sysml Block Visibility
