+++ src/main/java/org/sysml/ast/SysmlExpr.java
@@ -1,4 +1,4 @@
-package org.sysml;
+package org.sysml.ast;
 
 import java.io.IOException;
