+++ src/main/java/org/sysml/ast/SysmlPartDef.java
@@ -1,4 +1,4 @@
-package org.sysml;
+package org.sysml.ast;
 
 public class SysmlPartDef {
 }
